Completed JWT and bcrypt
