--- demo/src/error/badRequest.js
@@ -1,11 +1,0 @@
-import { StatusCodes } from "http-status-codes";
-import { CustomApiError } from "./customError.js";
-
-class BadRequestError extends CustomApiError {
-  constructor(message) {
-    super(message);
-    this.statusCode = StatusCodes.BAD_REQUEST;
-  }
-}
-
-export { BadRequestError };
