--- demo/src/error/customError.js
@@ -1,7 +1,0 @@
-class CustomApiError extends Error {
-  constructor(message) {
-    super(message);
-  }
-}
-
-export { CustomApiError };
