--- demo/src/error/index.js
@@ -1,5 +1,0 @@
-import { CustomApiError } from "./customError.js";
-import { BadRequestError } from "./badRequest.js";
-import { UnauthenticatedError } from "./unAuthenticated.js";
-
-export { CustomApiError, BadRequestError, UnauthenticatedError };
