--- demo/src/utils/commonSchema.js
@@ -1,6 +1,0 @@
-const commonSchema = {
-  created_at: { type: Date, default: Date.now() },
-  updated_at: { type: Date, default: Date.now() },
-};
-
-export { commonSchema };
